Adding literacy_groups feature with many changes
This commit contains big changes to dozens of files, implementing a
teacher mode on the site. Added the concept of literacy_groups, called
groups in the UI, which allows leaders (teachers) to create groups,
invite other users to join the groups, and assign prompts to those
users.

I reworked the existing prompts infrastructure so that prompts now
belong to groups. Instead of using the native
django.contrib.auth.models.Group, LiteracyGroups have foreign key
relationships for members and leaders. The previous notion of prompt
owners and prompt assignees is now mapped to group leaders and group
members respectively.